Reorganization into separate headers for helper, random functions.
